refactor: Move and rename taxonomy UI components to match 0002 ADR
